Crate dip_provider_runtime_template

Runtime template of a Decentralized Identity Provider (DIP) provider, which includes, beyond system pallets, did::Pallet, pallet_web3_names::Pallet, and pallet_did_lookup::Pallet pallets, as well as the pallet_dip_provider::Pallet pallet and the pallet_deposit_storage::Pallet pallet.

Modules

api

deposit

runtime_api

Macros

add_benchmarks

Calls add_benchmark with all configs from define_benchmarks and passes the first two parameters on.

list_benchmarks

Calls list_benchmark with all configs from define_benchmarks and passes the first two parameters on.

Structs

MaxNewKeyAgreementKeys

MaxTotalKeyAgreementKeys

PalletInfo

Provides an implementation of PalletInfo to provide information about the pallet setup in the runtime.

Perbill

A fixed point representation of a number in the range [0, 1].

Permill

A fixed point representation of a number in the range [0, 1].

PotId

RelayOrigin

Runtime

RuntimeApi

RuntimeApiImpl

Implements all runtime apis for the client side.

RuntimeBlockLength

RuntimeBlockWeights

RuntimeGenesisConfig

RuntimeOrigin

The runtime origin type representing the origin of a call.

SessionKeys

Version

Enums

MultiAddress

A multi-format address wrapper for on-chain accounts.

OriginCaller

RuntimeCall

RuntimeError

RuntimeEvent

RuntimeFreezeReason

A reason for placing a freeze on funds.

RuntimeHoldReason

A reason for placing a hold on funds.

RuntimeLockId

An identifier for each lock placed on funds.

RuntimeSlashReason

A reason for slashing funds.

RuntimeTask

An aggregation of all Task enums across all pallets included in the current runtime.

Constants

EXISTENTIAL_DEPOSIT

HOURS

MAX_PUBLIC_KEYS_PER_DID

MAX_REVEALABLE_LINKED_ACCOUNTS

MILLISECS_PER_BLOCK

MILLIUNIT

MINUTES

SS58_PREFIX

UNIT

VERSION

WASM_BINARY

WASM_BINARY_BLOATY

Traits

BuildStorage

Complex storage builder stuff.

Functions

native_version

Type Aliases

AccountId

Address

AllPalletsWithSystem

All pallets included in the runtime as a nested tuple of types.

AllPalletsWithoutSystem

All pallets included in the runtime as a nested tuple of types. Excludes the System pallet.

Aura

AuraConfig

AuraExt

AuraExtConfig

AuraId

An Aura authority identifier using S/R 25519 as its crypto.

Authorship

Balance

Balances

BalancesConfig

Block

BlockNumber

CheckedExtrinsic

CollatorSelection

CollatorSelectionConfig

DepositStorage

Did

DidIdentifier

DidLookup

DidLookupConfig

DipProvider

Executive

GenesisConfigDeprecated

Hash

Header

NodeBlock

Nonce

ParachainInfo

ParachainInfoConfig

ParachainSystem

ParachainSystemConfig

Session

SessionConfig

Signature

SignedExtra

Sudo

SudoConfig

System

SystemConfig

Timestamp

TransactionPayment

TransactionPaymentConfig

UncheckedExtrinsic

Utility

Web3Name

Web3Names